refact and improve TotalCrossSDKManager class
TotalCrossSDKDownloader was renamed to TotalCrossSDKManager, this is a more appropriate name. In addition, now TotalCrossSDKManager downloads SDKs to a more suitable place that depends on the OS, i.e., on mac it downloads to ~/Library/Application Support/TotalCross; on Windows. Also, now this class throws a TotalCross plugin exception if the version is not found on our online repository.