adds option to add external files to the application package (#12)
* adds option to add external files to the application package

this is a replacement to all.pkg. by using externalResources, you can specify which files out of java folder resources must be inside the package of the final application. Files listed inside externalResources won't be zipped inside the .tcz of the application, the will take place at the root of the application.